+++ tests/tests_cycle/test_cases.py
@@ -52,6 +52,7 @@ TEST_INVERSE = [
     (Cycle(1, 3, 4, 2), Cycle(2, 4, 3, 1)),
     (Cycle(2, 3, 1, 5, 4), Cycle(4, 5, 1, 3, 2)),
 ]
+TEST_IS_CONJUGATE_ERROR = [(Cycle(1, 2, 3), Cycle(1, 2, 3), NotImplementedError, "Method")]
 TEST_IS_DERANGEMENT = [
     (Cycle(1), False),
     (Cycle(13), False),