+++ tests/tests_cycle_decomposition/test_generic_methods.py
@@ -11,6 +11,7 @@ from tests.test_factory import (
     validate_support,
     validate_cycle_type,
     validate_equivalent,
+    validate_is_conjugate,
     validate_cycle_notation,
     validate_is_derangement,
     validate_cycle_decomposition,
@@ -26,8 +27,10 @@ from tests.tests_cycle_decomposition.test_cases import (
     TEST_SUPPORT,
     TEST_CYCLE_TYPE,
     TEST_EQUIVALENT,
+    TEST_IS_CONJUGATE,
     TEST_CYCLE_NOTATION,
     TEST_IS_DERANGEMENT,
+    TEST_IS_CONJUGATE_ERROR,
     TEST_CYCLE_DECOMPOSITION,
 )
 
@@ -70,6 +73,27 @@ def test_cycle_type(cycle_decomposition, expected_value) -> None:
 def test_inverse(cycle_decomposition, expected_value) -> None:
     """Tests for the method `inverse()`."""
     validate_inverse(item=cycle_decomposition, expected_value=expected_value)
+
+
+@pytest.mark.parametrize(
+    argnames="cycle_decomposition, other, expected_value",
+    argvalues=TEST_IS_CONJUGATE,
+    ids=[f"{p}.is_conjugate({o})={e}" for p, o, e in TEST_IS_CONJUGATE],
+)
+def test_is_conjugate(cycle_decomposition, other, expected_value) -> None:
+    """Tests for the method `is_derangement()`."""
+    validate_is_conjugate(item=cycle_decomposition, other=other, expected_value=expected_value)
+
+
+@pytest.mark.parametrize(
+    argnames="cycle_decomposition, other, error, msg",
+    argvalues=TEST_IS_CONJUGATE_ERROR,
+    ids=[f"{p}.is_conjugate({o})" for p, o, _, _ in TEST_IS_CONJUGATE_ERROR],
+)
+def test_is_conjugate_error(cycle_decomposition, other, error, msg) -> None:
+    """Tests for errors in the method `is_derangement()`."""
+    with pytest.raises(error, match=msg):
+        _ = cycle_decomposition.is_conjugate(other)
 
 
 @pytest.mark.parametrize(